feat: ContentView.swift 분리 및 ended 상태에서 시간이 정지되도록 수정
