+++ targets/watch/WorkoutUI.swift
@@ -17,15 +17,27 @@ final class WorkoutUI: ObservableObject {
   @Published var startedAt: Date? = nil          // 세션 시작 시각
   @Published var pauseAccum: TimeInterval = 0    // 누적 일시정지 시간
   @Published var pauseStartedAt: Date? = nil     // 현재 일시정지 시작 시각
+  @Published var endedAt: Date? = nil            // 종료 시각
 
   /// 지정 시점 기준 경과시간(초). paused이면 멈춘 값 유지.
   func elapsed(at date: Date = Date()) -> TimeInterval {
     guard let t0 = startedAt else { return 0 }
-    let now = (state == "paused") ? (pauseStartedAt ?? date) : date
+    let now: Date
+    switch state {
+    case "ended":
+        now = endedAt ?? date
+    case "paused":
+        now = pauseStartedAt ?? date
+    default:
+        now = date
+    }
+    
     var paused = pauseAccum
-    if let ps = pauseStartedAt, state != "running" {
+    
+    if let ps = pauseStartedAt, state != "running" && state != "ended" {
       paused += now.timeIntervalSince(ps)
     }
+    
     return max(0, now.timeIntervalSince(t0) - paused)
   }
 }
